feat[Atomic Design]: extract sign up form logic into AuthPagesSignUpForm organism
